implemented barebones sequential versions of map. need to figure out how to handle input arrays of different sizes and m's and n's.

+++ src/runtime/sequential.h
@@ -19,14 +19,14 @@ par_array seq_concat(const par_array a, const par_array b);
 par_array seq_select(const par_array a, int m, int n);
 
 // Maybe we can use variable arguments instead of defining functions for each num of input arrays (or maybe not seeing how the function to apply (f) needs to take the same amount of arguments as there are arrays passed, maybe a compiler thing to fix later on?)
-par_array seq_map1(float (*f)(float x), const par_array a);
-par_array seq_map2(float (*f)(float x, float y), const par_array a, const par_array b);
-par_array seq_map3(float (*f)(float x, float y, float z), const par_array a, const par_array b, const par_array c);
+par_array seq_map1(double (*f)(double x), const par_array a);
+par_array seq_map2(double (*f)(double x, double y), const par_array a, const par_array b);
+par_array seq_map3(double (*f)(double x, double y, double z), const par_array a, const par_array b, const par_array c);
 
-float seq_reduce(float (*f)(float x, float y), const par_array a);
-par_array seq_scan(float (*f)(float x, float y), const par_array a);
+double seq_reduce(double (*f)(double x, double y), const par_array a);
+par_array seq_scan(double (*f)(double x, double y), const par_array a);
 
 // Returns a new array where the elements fulfill predicate p, should probably be changed somehow to be more useful.
-par_array seq_mask(int (*p)(float x), par_array a);
+par_array seq_mask(int (*p)(double x), par_array a);
 
 #endif // SEQUENTIAL_H
